requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
numpy==1.21.2
tensorflow==2.6.0
pytorch==1.9.0
transformers==4.10.0
python-socketio==5.4.0
sqlalchemy==1.4.23
pillow==8.3.2
pandas==1.3.3
matplotlib==3.4.3
scikit-learn==0.24.2
redis==3.5.3
flask==3.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
werkzeug==3.0.1
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2